+++ spec/services/csv_import_spec.rb
@@ -320,7 +320,8 @@ describe CsvImport do
                                                                                                          'metro' => 'metro',
                                                                                                          'region' => 'region',
                                                                                                          'appeals' => 'appeals',
-                                                                                                         'website' => 'website')
+                                                                                                         'website' => 'website',
+                                                                                                         'referred_by' => 'referred_by')
     end
 
     it 'assigns file_constants when setting file' do
@@ -346,13 +347,13 @@ describe CsvImport do
         [['Johnny and Janey Doey', ' John', 'Doe', 'Jane ', 'Doe', 'Hi John and Jane', 'Doe family', 'Westside Baptist Church',
           '1 Example Ave, Apt 6', 'Sample City', 'IL', '60201', 'USA', 'Praying', '50', 'Monthly', 'CAD',
           'Both', 'christmas-card,      family', ' john@example.com ', ' jane@example.com ', '(213) 222-3333',
-          '(407) 555-6666', 'test notes', 'No', 'Yes', 'metro', 'region', 'Yes', 'http://www.john.doe'],
+          '(407) 555-6666', 'test notes', 'No', 'Yes', 'metro', 'region', 'Yes', 'http://www.john.doe', 'Mary Kim'],
          ['Bobby & Saray Parky', 'Bob', 'Park', 'Sara', 'Kim', 'Hello!', nil, nil, '123 Street West ', 'A Small Town', 'Quebec',
           'L8D 3B9 ', 'Canada', 'Praying and giving', '10', 'Monthly', nil, 'Both', 'bob', 'bob@park.com',
-          'sara@kim.com', '+12345678901', '+10987654321', nil, 'Yes', 'No', 'metro', 'region', 'No', 'website'],
+          'sara@kim.com', '+12345678901', '+10987654321', nil, 'Yes', 'No', 'metro', 'region', 'No', 'website', nil],
          ['Joey Many', 'Joe', 'Man', nil, nil, nil, nil, nil, 'Apartment, Unit 123', 'Big City', 'BC', nil, 'CA', 'Praying',
           nil, nil, nil, 'Both', nil, 'joe@inter.net', nil, '123.456.7890', nil, 'notes', nil, 'Yes', 'metro',
-          'region', 'Yes', 'website']]
+          'region', 'Yes', 'website', nil]]
       )
     end
   end
@@ -372,7 +373,7 @@ describe CsvImport do
       expect(csv_import.generate_csv_from_file_row_failures).to eq('Error Message,fullname,fname,lname,Spouse-fname,Spouse-lname,greeting,mailing-greeting,church,' \
                                                                    'street,city,province,zip-code,country,status,amount,frequency,currency,newsletter,tags,' \
                                                                    'email-address,Spouse-email-address,phone,Spouse-phone-number,extra-notes,skip,likely-giver' \
-                                                                   ",metro,region,appeals,website\n\"Validation failed: Email is invalid, Email is invalid\",Bob" \
+                                                                   ",metro,region,appeals,website,referred_by\n\"Validation failed: Email is invalid, Email is invalid\",Bob" \
                                                                    ',Park,Sara,Kim,Hello!,,,123 Street West,A Small Town,Quebec,L8D 3B9,Canada,Praying and giving' \
                                                                    ',10,Monthly,,Both,bob,this is not a valid email,this is also not a valid email,+12345678901' \
                                                                    ",+10987654321,,Yes,No,metro,region,No,website\n\"Validation failed: First name can't be blank, " \